AutomationDeviceFarm/AmazonDeviceFarm/appium_test_package/requirements.txt
appium-python-client
opencv-python-headless
pytesseract
pytest
pillow
numpy

